+++ src/ProtocolTranslator/DropRockDTO.h
@@ -7,7 +7,7 @@ class DropRockDTO: public ProtocolDTO {
 public:
     ~DropRockDTO() override;
 
-    int getClassId() const override;
+    int16_t getClassId() const override;
 };
