+++ src/ProtocolTranslator/JumpDTO.h
@@ -5,7 +5,7 @@
 
 class JumpDTO: public ProtocolDTO {
 public:
-    int getClassId() const override;
+    int16_t getClassId() const override;
 
     ~JumpDTO() override;
 };
