+++ src/ProtocolTranslator/LiftRockDTO.cpp
@@ -1,13 +1,13 @@
 #include "LiftRockDTO.h"
 
-LiftRockDTO::LiftRockDTO(const int &rock_id) : _rock_id(rock_id) { }
+LiftRockDTO::LiftRockDTO(const int16_t &rock_id) : _rock_id(rock_id) { }
 
 LiftRockDTO::~LiftRockDTO() = default;
 
-int LiftRockDTO::getClassId() const {
+int16_t LiftRockDTO::getClassId() const {
     return PROTOCOL_LIFT_ROCK;
 }
 
-const int LiftRockDTO::getRockId() const {
+const int16_t LiftRockDTO::getRockId() const {
     return _rock_id;
 }
