+++ src/ProtocolTranslator/MoveLeftDTO.h
@@ -5,7 +5,7 @@
 
 class MoveLeftDTO: public ProtocolDTO {
 public:
-    int getClassId() const override;
+    int16_t getClassId() const override;
 
     ~MoveLeftDTO() override;
 };
