+++ src/ProtocolTranslator/ProtocolDTO.h
@@ -7,7 +7,7 @@
 class ProtocolDTO {
 public:
     virtual ~ProtocolDTO() = default;
-    virtual int getClassId() const = 0;
+    virtual int16_t getClassId() const = 0;
 };
 
 #endif //PORTAL_PROTOCOLDTO_H
